Patching HTTP/OAuth/Provider/Request.php to handle queries that do not specify a realm.
In the current code, the very first parameter will always be ommitted.

Example:
Authorization: OAuth oauth_signature="xxxxxx", realm="some_realm"

The oauth_signature parameter will be ommitted in this query. Normally,
a realm would be sent as the first parameter, which is handled in the
code, and ommitted properly. However, the OAuth 1.0a Core specification
does not impose any ordering of query parameters, including the realm.
This means that any parameter should be allowed to come first.

This patch examines the $parts variable before handing it to the
omission condition, enabling the explode() statements (and therefore the
rest of the fuction) to do their job.